split the method into 3 parts, add comments
