+++ WritePropertiesToCSV/Program.cs
@@ -3,69 +3,17 @@ using System.Collections.Generic;
 using System.IO;
 using System.Reflection;
 using WritePropertiesToCSV.Entities;
+using System.Linq;
+using WritePropertiesToCSV.GetData;
 
 namespace WritePropertiesToCSV
 {
-    
-    public static class PersonList
-    {
-        public static List<Person> GetListPerson()
-        {
-            return new List<Person>
-            {
-                new Person
-                {
-                    Age = 24,
-                    EyeColor = "blue",
-                    Name = "Juarez Mayo",
-                    Gender = "male",
-                    Company = "MAGNEATO",
-                    Address= "284 Kansas Place, Beyerville, Pennsylvania, 5206",
-                    Salary = (decimal?) 345.6
-                },
-                new Person
-                {
-                    Age= 26,
-                    EyeColor= "green",
-                    Name= "Orr Love",
-                    Gender= "male",
-                    Company= "BULLZONE",
-                    Address= "893 Beaver Street, Johnsonburg, Nebraska, 503",
-                    Salary = (decimal?) 99.32
-                },
-                new Person
-                {
-                    Age= 32,
-                    EyeColor= "blue",
-                    Name= "Mccall Munoz",
-                    Gender= "male",
-                    Company= "DOGNOST",
-                    Address= "850 Mill Road, Chemung, Mississippi, 2962",
-                    Salary = (decimal?) 3000.89
-                },
-                new Person
-                {
-                    EyeColor= "green",
-                    Name= "Strong Downs",
-                    Gender= "male",
-                    Company= "BEADZZA",
-                    Address= "377 Homecrest Court, Tuskahoma, New Jersey, 3583"
-                },
-                new Person
-                {
-                    EyeColor= "brown",
-                    Name = "Sarah Pope",
-                    Gender = "female"
-                }
-            };
-        }
-    }
-
     class Program
     {
         static void Main(string[] args)
         {
             StartApp();
+
             Console.ReadLine();
         }
 
@@ -73,10 +21,12 @@ namespace WritePropertiesToCSV
         {
             //Create file
             string fileName = "PersonInfo.csv";
+
             CreateFile(fileName);
 
             //Get all properties from Person type
             List<PropertyInfo> declaredProperties = typeof(Person).GetTypeInfo().GetProperties().ToList();
+
             string[] propertiesFromUser = GetPropertiesFromUser(declaredProperties);
 
             //Data with filtered properties
@@ -90,6 +40,7 @@ namespace WritePropertiesToCSV
 
         private static void AddDataToCSVFile(string[] finish, string fileName)
         {
+            //add filtered properties with data to csv file
             foreach (var str in finish)
             {
                 AddRecord(str, fileName);
@@ -98,45 +49,76 @@ namespace WritePropertiesToCSV
 
         private static string[] GetFilteredPersonPersonProperties(List<Person> getListPerson, string[] propertiesFromUser, List<PropertyInfo> declaredProperties)
         {
-            string[] finish = new string[getListPerson.Count];
+            string[] allFilteredPropertiesWithValueForAllUser = new string[getListPerson.Count];
+
             for (int i = 0; i < getListPerson.Count; i++)
             {
                 string filterUserProperties = "";
+
                 for (int j = 0; j < propertiesFromUser.Length; j++)
                 {
-                    for (int k = 0; k < declaredProperties.Count; k++)
-                    {
-                        if (propertiesFromUser[j].Replace(" ", "").ToLower() == declaredProperties[k].Name.ToLower())
-                        {
-                            //Check value in property
-                            PropertyInfo prop = getListPerson[i].GetType().GetProperty(declaredProperties[k].Name, BindingFlags.Instance | BindingFlags.Public);
-                            var val = prop.GetValue(getListPerson[i], null) != null ? prop.GetValue(getListPerson[i], null).ToString() : null;
-
-                            if (val != null && val != "0")
-                            {
-                                //Add property with value to string
-                                filterUserProperties = filterUserProperties + $"{declaredProperties[k].Name}: " + typeof(Person).GetProperty(declaredProperties[k].Name).GetValue(getListPerson[i]).ToString() + ",";
-                                break;
-                            }
-                        }
-                    }
-
+                    filterUserProperties = WritePropertiesForOneUser(getListPerson[i], propertiesFromUser[j], declaredProperties);
                 }
-                finish[i] = filterUserProperties;
+
+                allFilteredPropertiesWithValueForAllUser[i] = filterUserProperties;
             }
-            return finish;
+
+            return allFilteredPropertiesWithValueForAllUser;
         }
 
+        public static string WritePropertiesForOneUser(Person person, string propertyFromUser, List<PropertyInfo> declaredProperties)
+        {
+            //create finall string with all filtered properties
+            string stringWithFilterProperties = "";
+
+            for (int k = 0; k < declaredProperties.Count; k++)
+            {
+                if (propertyFromUser.Replace(" ", "").ToLower() == declaredProperties[k].Name.ToLower())
+                {
+                    //Check value in property
+                    PropertyInfo property = person.GetType().GetProperty(declaredProperties[k].Name, BindingFlags.Instance | BindingFlags.Public);
+
+                    //Get value from this property if there is no data, assign null
+                    string valueInProperty = property.GetValue(person, null) != null ? property.GetValue(person, null).ToString() : null;
+
+                    //If value exist, add to to string with all filtered properti
+                    stringWithFilterProperties = stringWithFilterProperties + GetString(property, person, valueInProperty);
+                }
+            }
+
+            return stringWithFilterProperties;
+        }
+
+        static string GetString(PropertyInfo property, Person person, string valueInProperty)
+        {
+            //create new string
+            string stringWithFilteredPropertyForUser = "";
+
+            //check string - is null or == 0
+            if (valueInProperty != null && valueInProperty != "0")
+            {
+                //Add property with value to string
+                stringWithFilteredPropertyForUser = stringWithFilteredPropertyForUser + $"{property.Name}: " + typeof(Person).GetProperty(property.Name).GetValue(person).ToString() + ",";
+            }
+
+            return stringWithFilteredPropertyForUser;
+        }
 
         private static string[] GetPropertiesFromUser(IEnumerable<PropertyInfo> typeProperties)
         {
+            //Show user available properties
             Console.WriteLine("You can write only this properties: ");
+            
             foreach (var property in typeProperties)
             {
                 Console.WriteLine($"{property.Name}");
             }
+
             Console.Write("\nPlease, enter user properties: ");
+
+            //take properties from user
             string[] massOfpropertiesFromUser = Console.ReadLine().Split(',');
+
             return massOfpropertiesFromUser;
         }
 
@@ -144,6 +126,7 @@ namespace WritePropertiesToCSV
         {
             try
             {
+                //write to csv
                 using (StreamWriter file = new StreamWriter(path, true))
                 {
                     file.WriteLine(data);
@@ -157,12 +140,19 @@ namespace WritePropertiesToCSV
 
         public static void CreateFile(string path)
         {
+            //create file
             var file = new FileInfo(path);
+
             FileStream stream;
+
             if (!file.Exists)
+            {
                 stream = file.Create();
+            }
             else
+            {
                 return;
+            }
 
             stream.Close();
         }
